+++ src/index.ts
@@ -10,6 +10,7 @@ const app = express();
 config();
 
 /** Connect to Mongo */
+mongoose.set("strictQuery", false);
 mongoose
   .connect(process.env.MONGO_DB_URL || "", { retryWrites: true, w: "majority" })
   .then(() => {
