+++ src/models/product/index.ts
@@ -1,6 +1,4 @@
-import { Gender } from "./../user/index";
 import { Schema, model, Document } from "mongoose";
-import { type } from "os";
 
 /*********************TYPE & INTERFACE*****************************/
 export enum ProductSort {
@@ -14,7 +12,6 @@ export type ReviewType = {
   content: string;
   rating: number;
   created_at: string;
-  updated_at: string;
   user_id: string;
   username: string;
 };
@@ -24,10 +21,11 @@ export type ProductType = {
   price: number;
   description: string;
   colors: string[];
-  size: string[];
-  image: string;
-  category: string;
-  review: ReviewType;
+  sizes: string[];
+  banner: string;
+  images: string[];
+  categories: string[];
+  reviews: ReviewType[];
 };
 
 export type ProductTypeModel = {} & ProductType & Document;
@@ -36,7 +34,6 @@ export const Review = {
   content: String,
   rating: Number,
   created_at: String,
-  updated_at: String,
   user_id: String,
   username: String,
 };
@@ -45,11 +42,14 @@ const productSchema = new Schema({
   name: { type: String, require: true },
   price: { type: Number, require: true },
   description: { type: String, require: true },
-  colors: { type: String, default: [] },
-  size: { type: String, default: [] },
-  image: { type: String, require: true },
-  category: { type: String, require: true },
-  review: [Review],
+  colors: { type: [String], default: [] },
+  sizes: { type: [String], default: [] },
+  banner: { type: String, default: "" },
+  images: { type: [String], default: [] },
+  categories: { type: [String], default: [] },
+  reviews: { type: [Review], default: [] },
 });
 
-export default model<ProductTypeModel>("Product", productSchema);
+const Product = model<ProductTypeModel>("Product", productSchema);
+
+export default Product;
